requirements.txt
# Requirements for MLOps Vertex
google-cloud-aiplatform==1.7.1
google-cloud-pipeline-components==0.2.1
tensorflow==2.6.2
kfp==1.8.10
pre-commit
